test: 分割されていたテストファイルをまとめた
BM-control.test.tsxとBookmarkManager.test.tsxをBookmarkManager.test.tsxにまとめた。
